refactor(foundation-demo): pull shipping-options component out of shipping-form
pull shipping-options out of shipping-form
