+++ apps/foundation-demo/src/app/checkout/components/shipping/shipping-form/shipping-form.component.html
@@ -39,6 +39,6 @@
     name="telephone"
     [formControl]="form.controls['telephone']"
     [formSubmitted]="ngForm.submitted"/>
-  <shipping-options class="shipping-form__options" [selectedShippingOption]="selectedShippingOption" [shippingOptions]="shippingOptions" (selectShippingOption)="onSelectShippingOption($event)"></shipping-options>
+  <ng-content></ng-content>
   <button type="submit" class="button large expanded" [disabled]="!selectedShippingOption">{{hideContinueToPayment ? 'Save' : 'Continue to Payment'}}</button>
 </form>
